refactor:Extract stream manager as the public part
