+++ cri/stream/remotecommand/attach.go
@@ -1,6 +1,7 @@
 package remotecommand
 
 import (
+	gocontext "context"
 	"fmt"
 	"net/http"
 	"time"
@@ -9,12 +10,12 @@ import (
 // Attacher knows how to attach a running container in a pod.
 type Attacher interface {
 	// Attach attaches to the running container in the pod.
-	Attach(containerID string, streamOpts *Options, streams *Streams) error
+	Attach(goctx gocontext.Context, containerID string, streamOpts *Options, streams *Streams) error
 }
 
 // ServeAttach handles requests to attach to a container. After creating/receiving the required
 // streams, it delegates the actual attaching to attacher.
-func ServeAttach(w http.ResponseWriter, req *http.Request, attacher Attacher, container string, streamOpts *Options, idleTimeout, streamCreationTimeout time.Duration, supportedProtocols []string) {
+func ServeAttach(goctx gocontext.Context, w http.ResponseWriter, req *http.Request, attacher Attacher, container string, streamOpts *Options, idleTimeout, streamCreationTimeout time.Duration, supportedProtocols []string) {
 	ctx, ok := createStreams(w, req, streamOpts, supportedProtocols, idleTimeout, streamCreationTimeout)
 	if !ok {
 		// Error is handled by createStreams.
@@ -22,7 +23,7 @@ func ServeAttach(w http.ResponseWriter, req *http.Request, attacher Attacher, co
 	}
 	defer ctx.conn.Close()
 
-	err := attacher.Attach(container, streamOpts, &Streams{
+	err := attacher.Attach(goctx, container, streamOpts, &Streams{
 		StreamCh:     make(chan struct{}, 1),
 		StdinStream:  ctx.stdinStream,
 		StdoutStream: ctx.stdoutStream,
